wip: resign app validation

+++ packages/typescript-private/device-server/src/device-host/device-host.resign-app-file.ts
@@ -4,10 +4,10 @@ import { ChildProcess, HostPaths } from '@dogu-tech/node';
 import { Injectable } from '@nestjs/common';
 import AsyncLock from 'async-lock';
 import axios from 'axios';
-import compressing from 'compressing';
 import fs from 'fs';
 import path from 'path';
 import { v4 as uuidv4 } from 'uuid';
+import { env } from '../env';
 import { DoguLogger } from '../logger/logger';
 import { ResignShPath } from '../res-map';
 
@@ -35,28 +35,43 @@ export class DeviceHostResignAppFileService {
   }
 
   private async resign(message: ResignAppFileRequestBody): Promise<ResignAppFileResponseBodyData> {
+    const appPath = message.filePath;
     const doguHome = HostPaths.doguHomePath;
     const configsPath = HostPaths.configsPath(doguHome);
-    const certificateName = 'Apple Development: yow@dogutech.io (6KVR5VD4VM)';
+    const identityName = env.APPLE_RESIGN_IDENTITY_NAME;
     const provisioningProfilePath = HostPaths.resignProvisoningProfilePath(configsPath);
+
+    if (!appPath.endsWith('.ipa')) {
+      return { result: 'not-ipa' };
+    }
+
+    if (0 == identityName.length) {
+      return { result: 'no-identity' };
+    }
+    if (!fs.existsSync(provisioningProfilePath)) {
+      return { result: 'no-provisioning' };
+    }
 
     if (!fs.existsSync(HostPaths.doguTempPath())) {
       await fs.promises.mkdir(HostPaths.doguTempPath(), { recursive: true });
     }
 
     const contents = await fs.promises.readFile(ResignShPath, { encoding: 'utf-8' });
-    const shellPath = path.resolve(HostPaths.doguTempPath(), 'resign.sh');
+    const shellPath = path.resolve(HostPaths.doguTempPath(), `resign-${uuidv4()}.sh`);
     await fs.promises.writeFile(shellPath, contents, { encoding: 'utf-8' });
+    await fs.promises.chmod(shellPath, 0o755);
 
-    const appPath = message.filePath;
-    const tmpIpaPath = path.resolve(HostPaths.doguTempPath(), `resign-${uuidv4()}.zip`);
-    const tmpUnzipPath = path.resolve(HostPaths.doguTempPath(), `resign-${uuidv4()}-ipadir`);
-    const args = [appPath, `"${certificateName}"`, '-p', provisioningProfilePath, tmpIpaPath];
+    const resignedAppPath = path.resolve(HostPaths.doguTempPath(), `resign-${uuidv4()}.ipa`);
+    const args = [shellPath, appPath, `"${identityName}"`, '-p', provisioningProfilePath, resignedAppPath].join(' ');
 
-    await ChildProcess.spawnAndWait(shellPath, args, { shell: true, stdio: 'inherit' }, this.logger);
-    await compressing.zip.uncompress(tmpIpaPath, tmpUnzipPath);
+    this.logger.info('DeviceHostResignAppFileService.resign', { args });
+    const result = await ChildProcess.exec(args, {}, this.logger);
+    this.logger.info('DeviceHostResignAppFileService.resign done', { result });
     await fs.promises.rm(appPath, { recursive: true, force: true });
-    await fs.promises.cp(`${tmpUnzipPath}/Payload/${path.basename(appPath)}`, appPath, { recursive: true, force: true });
-    return {};
+    await fs.promises.cp(resignedAppPath, appPath, { recursive: true, force: true });
+
+    await fs.promises.rm(shellPath, { recursive: true, force: true });
+    await fs.promises.rm(resignedAppPath, { recursive: true, force: true });
+    return { result: 'success' };
   }
 }
